+++ Source/BIBD_Enumerator.h
@@ -35,9 +35,9 @@ protected:
 			i, j, lambda, this->getInSys()->lambda());
 	}
 	CK virtual const char *getObjName() const                       { return "BIBD"; }
-	CK virtual int addLambdaInfo(char *buffer, size_t lenBuffer, const char *pFrmt = NULL, int *pLambdaSetSize = NULL) const
+	CK virtual int addLambdaInfo(char *buffer, size_t lenBuffer, const char *pFrmt = NULL, size_t *pLambdaSetSize = NULL) const
 		{ return SNPRINTF(buffer, lenBuffer, pFrmt, lambda()); }
-	int addLambdaInform(const Class1(CVector)* lambdaSet, char* buffer, size_t lenBuffer, int *pLambdaSetSize) const;
+	int addLambdaInform(const Class1(CVector)* lambdaSet, char* buffer, size_t lenBuffer, size_t *pLambdaSetSize) const;
 	CK virtual void setFirstUnforcedRow(T rowNum = 0)				{}
 	CK virtual void resetFirstUnforcedRow()							{}
 	CK inline void setR(S val)										{ m_r = val; }
